Refactor and fix challenge caching
The previous code had a bunch of big tuples that were passed to
functions taking a lot of positional arguments with splats. This is not
ideal, especially when many of the arguments have the same types. This
commit refactors the challenge info into a dataclass, moves some caching
code into separate functions, and properly caches challenge info when
fetching all challenges.